SmoQyKPMCore

Documentation for SmoQyKPMCore.

SmoQyKPMCore.kpm_coefs

SmoQyKPMCore.kpm_coefs!

SmoQyKPMCore.kpm_coefs — Function

kpm_coefs(func::Function, bounds, M::Int, N::Int = 2*M)

Calculate and return the Chebyshev expansion coefficients. Refer to kpm_coefs! for more information.

SmoQyKPMCore.kpm_coefs! — Method

kpm_coefs!(
    coefs::AbstractVector{T}, func::Function, bounds,
    buf::Vector{T} = zeros(T, 2*length(coefs)),
    r2rplan::FFTW.r2rFFTWPlan = FFTW.plan_r2r!(zeros(T, 2*length(coefs)), FFTW.REDFT10)
) where {T<:AbstractFloat}

Calculate and record the Chebyshev expansion coefficients to order M in the vector ceofs for the function func on the bounded interval (bounds[1], bounds[2]). Let N be the number of evenly spaced points on the specified interval for which func is evaluated when performing Chebyshev-Gauss quadrature to compute the coefficients.