API

Note that for many methods there are two verions, one that relies on taking an instance of the KPMExpansion type as an argument, and a lower level one that does not.

KPMExpansion

update_kpmexpansion!

update_kpmexpansion_bounds!

update_kpmexpansion_order!

kpm_coefs

kpm_coefs!

kpm_mul

kpm_mul!

kpm_eval

kpm_eval!

apply_jackson_kernel

apply_jackson_kernel!

lanczos

lanczos!

SmoQyKPMCore.KPMExpansion — Type

mutable struct KPMExpansion{T<:AbstractFloat, Tfft<:FFTW.r2rFFTWPlan}

A type to represent the Chebyshev polynomial expansion used in the KPM algorithm.

Fields

M::Int: Order of the Chebyshev polynomial expansion.

bounds::NTuple{2,T}: Bounds on eigenspectrum in the KPM algorithm.

buf::Vector{T}: The first M elements of this vector of the Chebyshev expansion coefficients.

r2rplan::Tfft: Plan for performing DCT to efficiently calculate the Chebyshev expansion coefficients via Chebyshev-Gauss quadrature.

SmoQyKPMCore.KPMExpansion — Method

KPMExpansion(func::Function, bounds, M::Int, N::Int = 2*M)

Initialize an instance of the KPMExpansion type to approximate the univariate function func, called as func(x), with a order M Chebyshev polynomial expansion on the interval bounds[1] < x bounds[2]. Here, N ≥ M is the number of points at which func is evaluated on that specified interval, which are then used to calculate the expansion coeffiencents via Chebyshev-Gauss quadrature.

SmoQyKPMCore.update_kpmexpansion! — Function

update_kpmexpansion!(
    kpm_expansion::KPMExpansion{T}, func::Function, bounds, M::Int, N::Int = 2*M
) where {T<:AbstractFloat}

In-place update an instance of KPMExpansion to reflect new values for eigenspectrum bounds, expansion order M, and number of points at which the expanded function is evaluated when computing the expansion coefficients. This includes recomputing the expansion coefficients.

SmoQyKPMCore.update_kpmexpansion_bounds! — Function

update_kpmexpansion_bounds!(
    kpm_expansion::KPMExpansion{T}, func::Function, bounds
) where {T<:AbstractFloat}

In-place update an instance of KPMExpansion to reflect new values for eigenspectrum bounds, recomputing the expansion coefficients.

SmoQyKPMCore.update_kpmexpansion_order! — Function

update_kpmexpansion_order!(
    kpm_expansion::KPMExpansion, func::Function, M::Int, N::Int = 2*M
)

In-place update the expansion order M for an instance of KPMExpansion, recomputing the expansion coefficients. It is also possible to udpate the number of point N the function func is evaluated at to calculate the expansion coefficients.

SmoQyKPMCore.kpm_coefs — Function

kpm_coefs(func::Function, bounds, M::Int, N::Int = 2*M)

Calculate and return the Chebyshev expansion coefficients. Refer to kpm_coefs! for more information.

SmoQyKPMCore.kpm_coefs! — Function

kpm_coefs!(
    coefs::AbstractVector{T}, func::Function, bounds,
    buf::Vector{T} = zeros(T, 2*length(coefs)),
    r2rplan::FFTW.r2rFFTWPlan = FFTW.plan_r2r!(zeros(T, 2*length(coefs)), FFTW.REDFT10)
) where {T<:AbstractFloat}

Calculate and record the Chebyshev polynomial expansion coefficients to order M in the vector ceofs for the function func on the interval (bounds[1], bounds[2]). Let length(buf) be the number of evenly spaced points on the interval for which func is evaluated when performing Chebyshev-Gauss quadrature to compute the Chebyshev polynomial expansion coefficients.

SmoQyKPMCore.kpm_mul — Function

kpm_mul(
    A, coefs::AbstractVector, bounds, v::T,
    α₁::T = similar(v), α₂::T = similar(v), α₃::T = similar(v)
) where {T<:AbstractVecOrMat}

Evaluate and return the vector $v^\prime = F(A) \cdot v$ where $F(A)$ is represented by the Chebyshev expansion. For more information refer to kpm_mul!.

kpm_mul(A, kpm_expansion::KPMExpansion, v::T) where {T<:AbstractVector}

Evaluate and return the vector $v^\prime = F(A) \cdot v$ where $F(A)$ is represented by the Chebyshev expansion. For more information refer to kpm_mul!.

SmoQyKPMCore.kpm_mul! — Function

kpm_mul!(
    v′::T, A, coefs::AbstractVector, bounds, v::T,
    α₁::T = similar(v), α₂::T = similar(v), α₃::T = similar(v)
) where {T<:AbstractVecOrMat}

Evaluates $v^\prime = F(A) \cdot v$, writing the result to v′, where $F(A)$ is represented by the Chebyshev expansion. Here A is either a function that can be called as A(u,v) to evaluate $u = A\cdot v$, modifying u in-place, or is a type for which the operation mul!(u, A, v) is defined. The vector coefs contains Chebyshev expansion coefficients to approximate $F(A)$, where the eigenspectrum of $A$ is contained in the interval (bounds[1], bounds[2]) specified by the bounds argument. The vector v is vector getting multiplied by the Chebyshev expansion for $F(A)$. Lastly, the vectors (α₁, α₂, α₃) are passed to avoid dynamic memory allocations.

kpm_mul!(
    v′::T, A, kpm_expansion::KPMExpansion, v::T,
    α₁::T = similar(v), α₂::T = similar(v), α₃::T = similar(v)
) where {T<:AbstractVecOrMat}

Evaluates $v^\prime = F(A) \cdot v$, writing the result to v′, where $F(A)$ is represented by the Chebyshev expansion. Here A is either a function that can be called as A(u,v) to evaluate $u = A\cdot v$, modifying u in-place, or is a type for which the operation mul!(u, A, v) is defined. Lastly, the vectors (α₁, α₂, α₃) are passed to avoid dynamic memory allocations.

SmoQyKPMCore.kpm_eval — Function

kpm_eval(x::AbstractFloat, coefs, bounds)

Evaluate $F(x)$ where $x$ is real number in the interval bounds[1] < x < bound[2], and the function $F(\bullet)$ is represented by a Chebyshev expansion with coefficients given by the vector coefs.

kpm_eval(A::AbstractMatrix, coefs, bounds)

Evaluate and return the matrix $F(A),$ where $A$ is an operator with strictly real eigenvalues that fall in the interval (bounds[1], bounds[2]) specified by the bounds argument, and the function $F(\bullet)$ is represented by a Chebyshev expansion with coefficients given by the vector coefs.

kpm_eval(x::T, kpm_expansion::KPMExpansion{T}) where {T<:AbstractFloat}

Evaluate $F(x)$ where $x$ is real number in the interval bounds[1] < x < bound[2], and the function $F(\bullet)$ is represented by a Chebyshev expansion with coefficients given by the vector coefs.

kpm_eval(A::AbstractMatrix{T}, kpm_expansion::KPMExpansion{T}) where {T<:AbstractFloat}

Evaluate and return the matrix $F(A),$ where $A$ is an operator with strictly real eigenvalues and the function $F(\bullet)$ is represented by a Chebyshev expansion with coefficients given by the vector coefs.

SmoQyKPMCore.kpm_eval! — Function

kpm_eval!(
    F::T, A, coefs::AbstractVector, bounds,
    T₁::T = similar(F), T₂::T = similar(F), T₃::T = similar(F)
) where {T<:AbstractMatrix}

Evaluate and write the matrix $F(A)$ to F, where $A$ is an operator with strictly real eigenvalues that fall in the interval (bounds[1], bounds[2]) specified by the bounds argument, and the function $F(\bullet)$ is represented by a Chebyshev expansion with coefficients given by the vector coefs. Lastly, the matrices (T₁, T₂, T₃) are used to avoid dynamic memory allocations.

kpm_eval!(
    F::T, A, kpm_expansion::KPMExpansion,
    T₁::T = similar(F), T₂::T = similar(F), T₃::T = similar(F)
) where {T<:AbstractMatrix}

Evaluate and write the matrix $F(A)$ to F, where $A$ is an operator with strictly real eigenvalues and the function $F(\bullet)$ is represented by a Chebyshev expansion with coefficients given by the vector coefs. Lastly, the matrices (T₁, T₂, T₃) are used to avoid dynamic memory allocations.

SmoQyKPMCore.apply_jackson_kernel — Function

apply_jackson_kernel(coefs)

Return the Chebyshev expansion coefficients transformed by the Jackson kernel.

SmoQyKPMCore.apply_jackson_kernel! — Function

apply_jackson_kernel!(coefs)

Modify the Chebyshev expansion coefficients by applying the Jackson kernel to them.

apply_jackson_kernel!(kpm_expansion::KPMExpansion)

Modify the Chebyshev expansion coefficients by applying the Jackson kernel to them.

SmoQyKPMCore.lanczos — Function

lanczos(niters, v, A, S = I, rng = Random.default_rng())

Use niters Lanczos iterations to find a truncated tridiagonal representation of $A\cdot S$, up to similarity transformation. Here, $A$ is any Hermitian matrix, while $S$ is both Hermitian and positive definite. Traditional Lanczos uses the identity matrix, $S = I$. The extension to non-identity matrices $S$ is as follows: Each matrix-vector product $A\cdot v$ becomes $(A S) \cdot v$, and each vector inner product $w^\dagger \cdot v$ becomes $w^\dagger \cdot S \cdot v$. The implementation below follows Wikipedia, and is the most stable of the four variants considered by Paige [1]. This implementation introduces additional vector storage so that each Lanczos iteration requires only one matrix-vector multiplication for $A$ and $S$, respectively.

This function returns a SymTridiagonal matrix. Note that the eigmin and eigmax routines have specialized implementations for a SymTridiagonal matrix type.

SmoQyKPMCore.lanczos! — Function

lanczos!(
    αs::AbstractVector, βs::AbstractVector, v::AbstractVector,
    A, S = I,
    tmp::AbstractMatrix = zeros(length(v), 5),
    rng = Random.default_rng()
)

Use Lanczos iterations to find a truncated tridiagonal representation of $A\cdot S$, up to similarity transformation. Here, $A$ is any Hermitian matrix, while $S$ is both Hermitian and positive definite. Traditional Lanczos uses the identity matrix, $S = I$. The extension to non-identity matrices $S$ is as follows: Each matrix-vector product $A\cdot v$ becomes $(A S) \cdot v$, and each vector inner product $w^\dagger \cdot v$ becomes $w^\dagger \cdot S \cdot v$. The implementation below follows Wikipedia, and is the most stable of the four variants considered by Paige [1]. This implementation introduces additional vector storage so that each Lanczos iteration requires only one matrix-vector multiplication for $A$ and $S$, respectively.

The number of Lanczos iterations performed equals niters = length(αs), and niters - 1 == length(βs). This function returns a SymTridiagonal matrix based on the contents of the vectors αs and βs. Note that the eigmin and eigmax routines have specialized implementations for a SymTridiagonal matrix type.